Adding organization controller and minor refactoring

+++ src/main/java/com/sda/transporeon/finalconferenceroom/conference_room/model/ConferenceRoomDto.java
@@ -8,9 +8,9 @@ import lombok.NoArgsConstructor;
 public class ConferenceRoomDto {
 
     private String conferenceRoomName;
-    private Short level;
+    private Integer level;
     private Boolean availability;
-    private Short sittingPlaces;
-    private Short standingPlaces;
+    private Integer sittingPlaces;
+    private Integer standingPlaces;
     private String organizationName;
 }
